users_service/pyproject.toml
[tool.poetry]
name = "users"
version = "0.1.0"
description = "User coordination service for managing users, service linking, and cross-service data aggregation."
authors = ["Peleke <kraken@peleke.33mail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0.0"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = "^2.10.6"
python-multipart = "^0.0.20"
strawberry-graphql = {extras = ["debug-server"], version = "^0.268.1"}
python-dotenv = "^1.0.1"
httpx = "^0.28.1"
pydantic-settings = "^2.9.1"
psycopg2-binary = "^2.9.9"
supabase = "^2.9.1"
sqlalchemy = "^2.0.41"
asyncpg = "^0.30.0"
docker = "^7.1.0"
returns = "^0.25.0"
toolz = "^1.0.0"

[tool.poetry.scripts]
start = "users.web.server:start"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.26.0"
pytest-cov = "^4.1.0"
pytest = "^8.0.0"
black = "^24.1.1"
flake8 = "^7.0.0"
flake8-bugbear = "^24.1.1"
flake8-comprehensions = "^3.14.0"
flake8-simplify = "^0.21.0"
mypy = "^1.8.0"
isort = "^6.0.1"
httpx = "^0.28.1"
pre-commit = "^4.2.0"
schemathesis = "^3.26.1"
flake8-pyproject = "^1.2.3"
docker = "^7.1.0"

[tool.tox]
legacy_tox_ini = "tox.ini"

[tool.isort]
profile = "black"
skip = ["node_modules", ".git", "__pycache__", "*.ipynb", ".tox"]

[tool.black]
line-length = 120
exclude = '''
/(
    node_modules|
    \.git|
    __pycache__|
    .*\.ipynb|
    .tox
)/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov=users --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]
exclude = [".tox", "*.egg", "build", "data", ".venv", "tests"]
per-file-ignores = [
    "__init__.py:F401",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
files = "users"
exclude = ["tests/"] 